Refactor: 챌린지 제안하기 페이지, 디테일페이지 전체적으로 코드 리팩토링
