+++ src/components/idea-detail-page/review/useReviewGetPost.hook.ts
@@ -4,7 +4,6 @@ import { useInView } from 'react-intersection-observer';
 import { getIdeaCommentInfinite, postChallengeIdeaComment } from '@/app/api/idea-comments';
 
 export default function useReview(slug: string, userId: string | undefined, comment: string, setComment: React.Dispatch<React.SetStateAction<string>>) {
-  // 해당 포스트 댓글 데이터 get
   const {
     data: commentsData,
     isError: commentsError,
@@ -30,7 +29,6 @@ export default function useReview(slug: string, userId: string | undefined, comm
     },
   });
 
-  // 댓글 insert
   const queryClient = useQueryClient();
   const postMutation = useMutation(postChallengeIdeaComment, {
     onSuccess: () => {
