+++ src/components/idea-post-page/useImagePost.ts
@@ -13,15 +13,14 @@ export default function useImagePost(setImgFile: React.Dispatch<React.SetStateAc
   }, []);
 
   const handleChangeImg = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedFile = e.target.files?.[0]; // Access the selected file
+    const selectedFile = e.target.files?.[0];
     if (selectedFile) {
       imgUpload(selectedFile);
     }
   };
 
-  // Drag & Drop 사진 첨부 => DB state 할당, 미리보기 state 할당 함수 실행
   const onDrop = useCallback((acceptedFiles: File[]) => {
-    const [selectedFile] = acceptedFiles; // Access the selected file
+    const [selectedFile] = acceptedFiles;
     const imgExtension = selectedFile.name.split('.')[1];
     const extension = ['jpeg', 'jpg', 'png', 'GIF'].includes(imgExtension);
     if (extension) {
@@ -31,7 +30,6 @@ export default function useImagePost(setImgFile: React.Dispatch<React.SetStateAc
     }
   }, []);
 
-  // 첨부된 파일 읽고 DB state 할당, 미리보기 state 할당
   const imgUpload = (selectedFile: File) => {
     if (selectedFile) {
       setImgFile(selectedFile);
@@ -45,7 +43,6 @@ export default function useImagePost(setImgFile: React.Dispatch<React.SetStateAc
     }
   };
 
-  // Image 취소 버튼 click시 실행
   const handleCancelImg = () => {
     setImgFile(undefined);
     setPreviewImg(undefined);
